backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
supabase
httpx
openai
Pillow
playwright
requests
python-dotenv
google-cloud-storage>=2.10.0
langchain>=0.1.0
langchain-openai>=0.1.0
tqdm
google-generativeai>=0.7.0
